+++ tasks/database.php
@@ -13,9 +13,9 @@
 
 namespace Gaambo\DeployerWordpress\Tasks;
 
-use Gaambo\DeployerWordpress\Files;
-use Gaambo\DeployerWordpress\Localhost;
 use Gaambo\DeployerWordpress\WPCLI;
+use Gaambo\DeployerUtils\Files;
+use Gaambo\DeployerUtils\Localhost;
 
 use function Deployer\get;
 use function Deployer\has;
@@ -46,7 +46,11 @@ task('db:remote:backup', function () {
     $resolvedRemoteDumpPath = Files::resolvePath($remoteDumpPath, get('deploy_path'));
 
     run("mkdir -p $resolvedRemoteDumpPath");
-    WPCLI::runCommand("db export $resolvedRemoteDumpPath/$dumpFile --add-drop-table");
+    $remoteDumpFile = "$resolvedRemoteDumpPath/$dumpFile";
+    WPCLI::runCommand(
+        "db export $remoteDumpFile --add-drop-table",
+        runtimePaths: [$remoteDumpFile]
+    );
 
     Files::pullFile("$resolvedRemoteDumpPath/$dumpFile", "$resolvedLocalDumpPath/$dumpFile");
 })->desc('Create backup of remote database and download locally');
@@ -75,7 +79,8 @@ task('db:local:backup', function () {
     Localhost::run("mkdir -p $resolvedLocalDumpPath");
     WPCLI::runCommandLocally(
         "db export $resolvedLocalDumpPath/$dumpFile --add-drop-table",
-        Localhost::getConfig('current_path')
+        Localhost::getConfig('current_path'),
+        runtimePaths: ["$resolvedLocalDumpPath/$dumpFile"]
     );
 
     Files::pushFile("$resolvedLocalDumpPath/$dumpFile", "$resolvedRemoteDumpPath/$dumpFile");
@@ -104,7 +109,8 @@ task('db:remote:import', function () {
 
     $localUrl = Localhost::getConfig('public_url');
     $remoteUrl = get('public_url');
-    WPCLI::runCommand("db import $resolvedRemoteDumpPath/$dumpFile");
+    $remoteDumpFile = "$resolvedRemoteDumpPath/$dumpFile";
+    WPCLI::runCommand("db import $remoteDumpFile", runtimePaths: [$remoteDumpFile]);
 
     if (get('wp/multisite')) {
         WPCLI::runCommand("search-replace $localUrl $remoteUrl --network --all-tables");
@@ -151,7 +157,8 @@ task('db:local:import', function () {
     $remoteUrl = get('public_url');
     WPCLI::runCommandLocally(
         "db import $resolvedLocalDumpPath/$dumpFile",
-        Localhost::getConfig('current_path')
+        Localhost::getConfig('current_path'),
+        runtimePaths: ["$resolvedLocalDumpPath/$dumpFile"]
     );
 
     if (get('wp/multisite')) {